Added new delta QC test plus updates to make tests more flexible with irregular timestamps.
Added QC test to check bounds on the difference between max and min
within a rolling window. Tests and documentation added. Note, this
method is currently NOT efficient for large data sets (> 100000 pts)
because it uses df.rolling().apply() to find the index of min and max
and to work with NaNs).  Updated rolling windows to accept time offset
instead of number of timesteps. Bug fix in check_timestamp, included NaN
in test. Updated input to append_test_results to clarify when the mask
is used directly to extract test results (used for timestamp tests, but
could be used for other applications).